src/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/custom/upgrade-candidate-adopt-crash-expected.txt

CONSOLE ERROR: line 148: Uncaught TypeError: text.replace is not a function
Tests that an element is not upgraded after being moved to a different document, then destroyed

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

FAIL Unexpected error: Uncaught TypeError: text.replace is not a function
FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
